+++ src/generator_gemm_amx_microkernel.h
@@ -38,7 +38,7 @@ void libxsmm_generator_gemm_amx_prefetch_tile_in_L2(libxsmm_generated_code*     
     unsigned int tile_cols,
     unsigned int LD,
     unsigned int base_reg,
-    unsigned int offset);
+    long long    offset);
 
 LIBXSMM_API_INTERN
 void libxsmm_generator_gemm_amx_paired_tilestore( libxsmm_generated_code*            io_generated_code,
@@ -66,7 +66,7 @@ void libxsmm_generator_gemm_amx_decompress_32x32_A_block(libxsmm_generated_code*
     libxsmm_loop_label_tracker*        io_loop_label_tracker,
     const libxsmm_gp_reg_mapping*      i_gp_reg_mapping,
     libxsmm_micro_kernel_config*       i_micro_kernel_config,
-    int                                a_offs,
+    long long                          a_offs,
     unsigned int                       a_lookahead_offs,
     unsigned int                       a_lookahead_br_index);
 
@@ -87,8 +87,8 @@ void libxsmm_generator_gemm_amx_microkernel( libxsmm_generated_code*            
                                                      const libxsmm_gemm_descriptor*     i_xgemm_desc,
                                                      libxsmm_blocking_info_t*           n_blocking_info,
                                                      libxsmm_blocking_info_t*           m_blocking_info,
-                                                     unsigned long long                 offset_A,
-                                                     unsigned long long                 offset_B,
+                                                     long long                          offset_A,
+                                                     long long                          offset_B,
                                                      unsigned int                       is_last_k,
                                                      long long                          i_brgemm_loop,
                                                      unsigned int                       fully_unrolled_brloop  );
@@ -101,8 +101,8 @@ void libxsmm_generator_gemm_amx_kernel_kloop( libxsmm_generated_code*           
                                                       const libxsmm_gemm_descriptor*     i_xgemm_desc,
                                                       libxsmm_blocking_info_t*           n_blocking_info,
                                                       libxsmm_blocking_info_t*           m_blocking_info,
-                                                      unsigned long long                 A_offs,
-                                                      unsigned long long                 B_offs,
+                                                      long long                          A_offs,
+                                                      long long                          B_offs,
                                                       unsigned int                       fully_unrolled_brloop );
 
 #endif /* GENERATOR_GEMM_AMX_MICROKERNEL_H */
